TS fixes js_only
