+++ assets/prototype/domain/eventEditor/datetimes/dateCard/styles.ts
@@ -1,0 +1,21 @@
+import { CSSProperties } from 'react';
+
+export const btnStyle: CSSProperties = {
+	margin: '0 0 0 .5rem',
+};
+
+export const cardStyle: CSSProperties = {
+	margin: '0 0 2rem',
+	minWidth: '360px',
+	position: 'relative' as 'relative', // cast value to type
+	textAlign: 'center' as 'center',
+	width: '32%',
+};
+
+export const idStyle: CSSProperties = {
+	color: 'grey',
+	fontSize: '9px',
+	left: '.75em',
+	position: 'absolute' as 'absolute',
+	top: '.5em',
+};
