+++ assets/prototype/domain/eventEditor/datetimes/types.ts
@@ -1,31 +1,18 @@
 import { Ticket, EntityId } from '../data/types';
-import { EditItemModalProps, EditItemFormProps, ListItemProps } from '../types';
-import { DatesSortedBy, DisplayDates, ShowDates } from '../data/date/types';
 
-export interface EditDateProps extends ListItemProps {
+export interface EditDateProps {
+	id: EntityId;
 	position?: string;
 	relatedTickets?: EntityId[];
 }
-export interface EditDateModalProps extends EditItemModalProps {
+export interface EditDateModalProps {
+	id: EntityId;
 	relatedTickets: EntityId[];
 }
 
-export interface DateItemFormProps extends EditItemFormProps {
+export interface DateItemFormProps {
 	tickets: Ticket[];
 	relatedTickets: EntityId[];
-}
-
-export interface DatesSortedByControlProps {
-	datesSortedBy: DatesSortedBy;
-	setDatesSortedBy: (datesSortedBy: DatesSortedBy) => void;
-}
-
-export interface DisplayDatesControlProps {
-	displayDates: DisplayDates;
-	setDisplayDates: (displayDates: DisplayDates) => void;
-}
-
-export interface ShowDatesControlProps {
-	showDates: ShowDates;
-	setShowDates: (showDates: ShowDates) => void;
+	title?: string;
+	formReset?: boolean;
 }
